# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Angel API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface UserProfile {
  id: number;
  name: string;
  sobrietyStart: string;
  addictionType?: string | null;
  dailySpend?: number | null;
  checkInTime?: string | null;
  createdAt: string;
}

export interface UpsertUserProfileBody {
  name: string;
  sobrietyStart: string;
  addictionType?: string | null;
  dailySpend?: number | null;
  checkInTime?: string | null;
}

export interface Guardian {
  id: number;
  userId: number;
  name: string;
  phone: string;
  relationship?: string | null;
  createdAt: string;
}

export interface CreateGuardianBody {
  userId: number;
  name: string;
  phone: string;
  relationship?: string | null;
}

export interface MoodLog {
  id: number;
  userId: number;
  moodScore: number;
  note?: string | null;
  createdAt: string;
}

export interface CreateMoodLogBody {
  userId: number;
  moodScore: number;
  note?: string | null;
}

export interface BuddyConnection {
  id: number;
  userId: number;
  sponsorId?: number | null;
  code: string;
  createdAt: string;
}

export interface CreateBuddyConnectionBody {
  userId: number;
  code: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type GeminiMessageRole =
  (typeof GeminiMessageRole)[keyof typeof GeminiMessageRole];

export const GeminiMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: GeminiMessageRole;
  content: string;
  createdAt: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: GeminiMessage[];
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export interface GeminiError {
  error: string;
}

export type ListMoodLogsParams = {
  days?: number;
};