# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Angel API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get user profile
 */
export const GetUserProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  sobrietyStart: zod.date(),
  addictionType: zod.string().nullish(),
  dailySpend: zod.number().nullish(),
  checkInTime: zod.string().nullish(),
  createdAt: zod.date(),
});

/**
 * @summary Create or update user profile
 */
export const UpsertUserProfileBody = zod.object({
  name: zod.string(),
  sobrietyStart: zod.date(),
  addictionType: zod.string().nullish(),
  dailySpend: zod.number().nullish(),
  checkInTime: zod.string().nullish(),
});

export const UpsertUserProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  sobrietyStart: zod.date(),
  addictionType: zod.string().nullish(),
  dailySpend: zod.number().nullish(),
  checkInTime: zod.string().nullish(),
  createdAt: zod.date(),
});

/**
 * @summary List guardian contacts
 */
export const ListGuardiansResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  relationship: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListGuardiansResponse = zod.array(ListGuardiansResponseItem);

/**
 * @summary Add a guardian contact
 */
export const CreateGuardianBody = zod.object({
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  relationship: zod.string().nullish(),
});

/**
 * @summary Remove a guardian
 */
export const DeleteGuardianParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List mood logs
 */
export const listMoodLogsQueryDaysDefault = 30;

export const ListMoodLogsQueryParams = zod.object({
  days: zod.coerce.number().default(listMoodLogsQueryDaysDefault),
});

export const ListMoodLogsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  moodScore: zod.number(),
  note: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListMoodLogsResponse = zod.array(ListMoodLogsResponseItem);

/**
 * @summary Log a mood entry
 */
export const CreateMoodLogBody = zod.object({
  userId: zod.number(),
  moodScore: zod.number(),
  note: zod.string().nullish(),
});

/**
 * @summary Get buddy connection info
 */
export const GetBuddyConnectionResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  sponsorId: zod.number().nullish(),
  code: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary Connect with a sponsor via code
 */
export const CreateBuddyConnectionBody = zod.object({
  userId: zod.number(),
  code: zod.string(),
});

export const CreateBuddyConnectionResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  sponsorId: zod.number().nullish(),
  code: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary List all conversations
 */
export const ListGeminiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListGeminiConversationsResponse = zod.array(
  ListGeminiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateGeminiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGeminiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListGeminiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListGeminiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListGeminiMessagesResponse = zod.array(
  ListGeminiMessagesResponseItem,
);

/**
 * @summary Send a message and get a streaming response
 */
export const SendGeminiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendGeminiMessageBody = zod.object({
  content: zod.string(),
});