# lib/api-zod/src/generated/types/geminiMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Angel API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiMessageRole } from "./geminiMessageRole";

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: GeminiMessageRole;
  content: string;
  createdAt: Date;
}