# lib/api-zod/src/generated/types/guardian.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Angel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Guardian {
  id: number;
  userId: number;
  name: string;
  phone: string;
  relationship?: string | null;
  createdAt: Date;
}