# lib/api-zod/src/generated/types/upsertUserProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Angel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertUserProfileBody {
  name: string;
  sobrietyStart: Date;
  addictionType?: string | null;
  dailySpend?: number | null;
  checkInTime?: string | null;
}